Strength Evaluator, List and Sort Strengths, Last Accessed List
re-implemented a strength evalulator for passwords which shows strengths in a list of passwords and updates the strength if the password is updated, implemented a sorting and listing by strengths for passwords, implemented a last accessed password list with the latest accessed password shown first with it's date and time. corrected code. added comments documentation.